11/14 회원 보기시 권한 검증 하기

+++ src/main/java/com/example/prj1be/controller/MemberController.java
@@ -64,9 +64,15 @@ public class MemberController {
     }
 
     @GetMapping
-    public ResponseEntity<Member> view(String id) {
-        // TODO : 로그인 했는 지? -> 안했으면 401
-        // TODO : 자기 정보인지? -> 아니면 403
+    public ResponseEntity<Member> view(String id,
+                                       @SessionAttribute(value = "login", required = false)Member login) {
+        if(login == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+        if(!service.hasAccess(id, login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
 
         Member member = service.getMember(id);
 
@@ -74,9 +80,15 @@ public class MemberController {
     }
 
     @DeleteMapping
-    public ResponseEntity delete(String id) {
-        // TODO : 로그인 했는 지? -> 안했으면 401
-        // TODO : 자기 정보인지? -> 아니면 403
+    public ResponseEntity delete(String id,
+                                 @SessionAttribute(value = "login", required = false)Member login) {
+        if(login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+        if(!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
 
         if (service.deleteMember(id)) {
             return ResponseEntity.ok().build();
@@ -87,8 +99,17 @@ public class MemberController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Member member) {
-        // TODO: 로그인 했는지? 자기정보인지?
+    public ResponseEntity edit(@RequestBody Member member,
+                               @SessionAttribute(value = "login", required = false) Member login) {
+
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+        if (!service.hasAccess(member.getId(), login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
+
 
         if (service.update(member)) {
             return ResponseEntity.ok().build();
